Add {client,server}_bridge_fn
Add the variables client_bridge_fn and server_bridge_fn that will be
called in bridge instead of always calling bridge_client and
bridge_server.

This will be used for those times where bridging is not the correct
action to take when data is available, like during an TLS/SSL handshake.