Make data selection use probability
Previously data handlers were all or nothing, but that doesn't work
well. Now we have passive handlers being run 100% of the time, since
they do nothing potentially destructive to a connection. For active
handlers we use the same probability as for connections and with
probability p we use all the selected active handlers, with 1-p we use
none of them.

We might later expand this to have per handler probability/selection,
but that gets complicated fast.

Fixes #20