Add Android app to test nogotofail MiTM.
This adds an Android application which offers an easy way to generate
vulnerable traffic which should be detected by the MiTM. This should
be useful for quickly testing the various attack/detection
functionality of the nogotofail MiTM.

This commit adds a small number of tests to cover the most typical
vulnerable traffic: HTTPS requests where the identity of the server is
not fully verified, and cleartext HTTP requests with an Authorizaton
header which thus leaks authentication credentials. More tests will be
added later.

Fixes #21