Use nonblocking connects when connecting to remote
The connection to the remote endpoint of a connection is now done
nonblocking instead of using a timeout.

This greately increases parallelism during connection creation since
applications tend to create multiple connections at the same time.

In addition we will no longer block for a second when connecting to an
endpoint that times out, though this happens very rarely.