Always use nonblocking socket operations.
With SSL wrapped sockets select is not a perfect indicator of if there
is application data ready for reading, so keep connection sockets nonblocking at
all times.

This also fixes issues in nogotofail.mitm.connection where socket.sendall could
block because of a full send buffer.

This temporarily disables SSLstrip/Webview JS RCE from loading as
default since they are especially fragile when it comes to short reads
of application data, which happens a lot more now. A fix will be
upcoming.