Add a check for connection loops
In tproxy and redirect mode clients connecting to the MiTM socket
directly instead of being routed through the iptables rules will cause
the MiTM to see the remote server address as itself, causing it to open
another connection to the MiTM and repeat until all fds are gone and
nogotofail crashes. If we detect such a client connection we will log a
warning and close that connection.