Add TlsRecordHandler base data handler
TlsRecordHandler handles buffering of TlsRecords in a connection,
handlers that want to do things based on records (eithe passive or
replacing records) can extend this class and implement on_tls_request
and on_tls_response.

In active TlsRecordHandlers these two methods return the _bytes_ to sent
on the wire in the place of the provided record. This allows you to send
completely different records, inject your own,  or modify the incomin
record arbitrarily.